IsBlankText rename to IsBlank
